+++ OggConverter/src/Forms/SettingsForm.Designer.cs
@@ -71,6 +71,8 @@
             this.btnOpenHistory = new System.Windows.Forms.Button();
             this.chkHistory = new System.Windows.Forms.CheckBox();
             this.chkCrashLog = new System.Windows.Forms.CheckBox();
+            this.label7 = new System.Windows.Forms.Label();
+            this.chkMono = new System.Windows.Forms.CheckBox();
             this.tabControl.SuspendLayout();
             this.tabGeneral.SuspendLayout();
             this.tabFiles.SuspendLayout();
@@ -192,6 +194,8 @@
             // 
             // tabFiles
             // 
+            this.tabFiles.Controls.Add(this.chkMono);
+            this.tabFiles.Controls.Add(this.label7);
             this.tabFiles.Controls.Add(this.chkRecommendedFrequency);
             this.tabFiles.Controls.Add(this.radQualityCompressed);
             this.tabFiles.Controls.Add(this.label6);
@@ -211,7 +215,7 @@
             // chkRecommendedFrequency
             // 
             this.chkRecommendedFrequency.AutoSize = true;
-            this.chkRecommendedFrequency.Location = new System.Drawing.Point(29, 116);
+            this.chkRecommendedFrequency.Location = new System.Drawing.Point(29, 292);
             this.chkRecommendedFrequency.Name = "chkRecommendedFrequency";
             this.chkRecommendedFrequency.Size = new System.Drawing.Size(413, 21);
             this.chkRecommendedFrequency.TabIndex = 27;
@@ -222,7 +226,7 @@
             // radQualityCompressed
             // 
             this.radQualityCompressed.AutoSize = true;
-            this.radQualityCompressed.Location = new System.Drawing.Point(29, 231);
+            this.radQualityCompressed.Location = new System.Drawing.Point(29, 210);
             this.radQualityCompressed.Name = "radQualityCompressed";
             this.radQualityCompressed.Size = new System.Drawing.Size(108, 21);
             this.radQualityCompressed.TabIndex = 7;
@@ -234,7 +238,7 @@
             // label6
             // 
             this.label6.AutoSize = true;
-            this.label6.Location = new System.Drawing.Point(26, 157);
+            this.label6.Location = new System.Drawing.Point(26, 136);
             this.label6.Name = "label6";
             this.label6.Size = new System.Drawing.Size(162, 17);
             this.label6.TabIndex = 6;
@@ -243,7 +247,7 @@
             // radQualityAverage
             // 
             this.radQualityAverage.AutoSize = true;
-            this.radQualityAverage.Location = new System.Drawing.Point(29, 204);
+            this.radQualityAverage.Location = new System.Drawing.Point(29, 183);
             this.radQualityAverage.Name = "radQualityAverage";
             this.radQualityAverage.Size = new System.Drawing.Size(82, 21);
             this.radQualityAverage.TabIndex = 5;
@@ -255,7 +259,7 @@
             // radQualityBest
             // 
             this.radQualityBest.AutoSize = true;
-            this.radQualityBest.Location = new System.Drawing.Point(29, 177);
+            this.radQualityBest.Location = new System.Drawing.Point(29, 156);
             this.radQualityBest.Name = "radQualityBest";
             this.radQualityBest.Size = new System.Drawing.Size(57, 21);
             this.radQualityBest.TabIndex = 4;
@@ -291,9 +295,9 @@
             this.chkAutoSort.AutoSize = true;
             this.chkAutoSort.Location = new System.Drawing.Point(29, 35);
             this.chkAutoSort.Name = "chkAutoSort";
-            this.chkAutoSort.Size = new System.Drawing.Size(191, 21);
+            this.chkAutoSort.Size = new System.Drawing.Size(240, 21);
             this.chkAutoSort.TabIndex = 1;
-            this.chkAutoSort.Text = "Sort files after conversion";
+            this.chkAutoSort.Text = "Automatically rearrange file order";
             this.chkAutoSort.UseVisualStyleBackColor = true;
             this.chkAutoSort.Click += new System.EventHandler(this.ChkAutoSort_Click);
             // 
@@ -551,6 +555,26 @@
             this.chkCrashLog.UseVisualStyleBackColor = true;
             this.chkCrashLog.Click += new System.EventHandler(this.ChkCrashLog_Click);
             // 
+            // label7
+            // 
+            this.label7.AutoSize = true;
+            this.label7.Location = new System.Drawing.Point(26, 272);
+            this.label7.Name = "label7";
+            this.label7.Size = new System.Drawing.Size(83, 17);
+            this.label7.TabIndex = 28;
+            this.label7.Text = "Conversion:";
+            // 
+            // chkMono
+            // 
+            this.chkMono.AutoSize = true;
+            this.chkMono.Location = new System.Drawing.Point(29, 319);
+            this.chkMono.Name = "chkMono";
+            this.chkMono.Size = new System.Drawing.Size(223, 21);
+            this.chkMono.TabIndex = 29;
+            this.chkMono.Text = "Convert song to mono channel";
+            this.chkMono.UseVisualStyleBackColor = true;
+            this.chkMono.Click += new System.EventHandler(this.chkMono_Click);
+            // 
             // SettingsForm
             // 
             this.AutoScaleDimensions = new System.Drawing.SizeF(8F, 16F);
@@ -619,5 +643,7 @@
         private System.Windows.Forms.TextBox txtAudacity;
         private System.Windows.Forms.RichTextBox txtChangelog;
         private System.Windows.Forms.CheckBox chkRecommendedFrequency;
+        private System.Windows.Forms.CheckBox chkMono;
+        private System.Windows.Forms.Label label7;
     }
 }